--- src/Controllers/article.php
@@ -1,25 +1,0 @@
-<?php
-
-namespace App\Controllers;
-
-use App\model\ArticleRepository;
-use App\model\DatabaseConnection;
-use Twig\Environment;
-
-class Article
-{
-    public function __construct(public Environment $twig)
-    {
-        
-    }
-
-    public function showArticle($id)
-    {
-        $id = $_GET['id'];
-        $articleRepository = new ArticleRepository();
-        $articleRepository->connection = new DatabaseConnection();
-        $article = $articleRepository->getArticle($id);
-
-        $this->twig->display('article.html.twig', ['article' => $article]);
-    }
-}
